Extract createLiteApi from the preload
The preload already built the renderer's api generically from the
endpoint lists; the construction is now a module over an abstract
transport, and the preload is its electron binding.

- LiteApiTransport is what a host provides: one request/response
  invoke, one named event channel with unsubscribe, and the platform.
  Electron backs them with ipcRenderer; another host backs them with
  its own channel and gets the identical, fully-typed surface
- createLiteApi builds every forwarded member from the lists, so a
  partial api is impossible — a missing member is a compile error, the
  exact failure a hand-written shim's cast once hid — and expresses
  the special members (askpass and full-screen events, AI streaming,
  the watcher lifecycle) over the same two primitives
- The watcher bookkeeping shrinks to a map of unsubscribe functions:
  channels stay inside the transport, where they belong
- preload.cts is down to the transport itself: ipcRenderer narrowed
  from Promise<any> to unknown, plus contextBridge exposure

No behavior change.
